requirements.txt
# Dependencies cho Tiki Data Pipeline

# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Environment variables
python-dotenv>=1.0.0

# Optional: For type checking (development)
# mypy>=1.0.0
# types-requests>=2.31.0

# Optional: For code formatting (development)
# black>=23.0.0
# ruff>=0.1.0
